Redirect large allocations to use malloc(), and not the shmem allocation.
We currently do not support free() for shared memory based allocation.
Some sub-systems that allocate and free large chunks repeatedly use
up too much shared memory, causing tests to fail.

This commit fixes that issue by re-directing following clients to
fall-back to using regular malloc() / free() interfaces:

  - merge_iterator_create()
  - merge_iterator_destroy(),
  - trunk_range() - trunk_range_iterator
  - trunk_range_iterator_deinit()
  - btree_pack_req_init()
  - btree_pack_req_deinit()

With these changes, unit-test splinter_test now succeeds using shared memory:

sdb-fdb-build:[10] $ build/debug/bin/unit/splinter_test
Running 1 CTests, suite name 'splinter', test case 'all'.
TEST 1/3 splinter:test_inserts [OK]
TEST 2/3 splinter:test_lookups [OK]
TEST 3/3 splinter:test_splinter_print_diags [OK]
RESULTS: 3 tests (3 ok, 0 failed, 0 skipped) ran in 1136485 ms

+++ src/platform_linux/platform.c
@@ -28,7 +28,7 @@ void __attribute__((constructor)) platform_init_log_file_handles(void)
 
 platform_status
 platform_heap_create(platform_module_id    UNUSED_PARAM(module_id),
-                     uint32                max,
+                     size_t                max,
                      bool                  use_shmem,
                      platform_heap_handle *heap_handle,
                      platform_heap_id     *heap_id)
